quart-telegram/requirements.txt
aiofiles==23.2.1
blinker==1.7.0
click==8.1.7
Flask==3.0.0
greenlet==3.0.3
h11==0.14.0
h2==4.1.0
hpack==4.0.0
Hypercorn==0.16.0
hyperframe==6.0.1
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.3
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==23.2
pathspec==0.12.1
pillow==10.2.0
platformdirs==4.1.0
priority==2.0.0
pyaes==1.6.1
pyasn1==0.5.1
pycodestyle==2.11.1
pyflakes==3.2.0
pypng==0.20220715.0
python-dotenv==1.0.0
qrcode==7.4.2
Quart==0.19.4
quart-auth==0.9.0
Quart-Session==3.0.0
redis==5.0.1
rsa==4.9
SQLAlchemy==2.0.25
Telethon==1.33.1
typing_extensions==4.9.0
Werkzeug==3.0.1
wsproto==1.2.0
